+++ src/components/RichTextEditor/Plugins/shared/alignment.ts
