+++ src/components/RichTextEditor/Plugins/shared/index.ts
@@ -1,3 +1,3 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
-export * from './textStyles';
+export * from './alignment';
